Handled problems of insertion of employees  and deleted of offices

+++ src/main/java/com/example/calender/service/OfficeServiceImpl.java
@@ -4,17 +4,23 @@ import com.example.calender.repository.OfficeRepository;
 import com.example.calender.entity.Office;
 import com.example.calender.exception.ResourceAlreadyExistsException;
 import com.example.calender.exception.ResourceNotFoundException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
-
+@Slf4j
 @Service
 public class OfficeServiceImpl implements OfficeService<Office>{
 
     @Autowired
     OfficeRepository officeRepository;
+    @Autowired
+    EmployeeService employeeService;
+    @Autowired
+    MeetingRoomService meetingRoomService;
     @Override
     public List<Office> getAllBranches() {return officeRepository.findAll();}
     @Override
@@ -32,10 +38,19 @@ public class OfficeServiceImpl implements OfficeService<Office>{
         //throw  new ResourceAlreadyExistsException("office","officeID",office.getOfficeID());
     }
 
+
+
     @Override
-    public void deleteOffice(long id) throws ResourceNotFoundException{
+    public boolean deleteOffice(long id) throws ResourceNotFoundException{
         officeRepository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Office","id",id));
-        officeRepository.deleteById(id);
+        if(employeeService.checkEmptyOffice(id)) {
+            officeRepository.deleteById(id);
+            List<Long> roomsInOffice = meetingRoomService.getMeetingRoomsByOfficeId(id);
+            for (Long aRoomId : roomsInOffice) meetingRoomService.deleteMeetingRoom(aRoomId);
+            return true;
+        }
+        else return false;
+
     }
 
     @Override
